requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
defusedxml==0.7.1
Django==4.1.4
django-allauth==0.51.0
django-filter==22.1
django-rest-auth-forked==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.14.0
exceptiongroup==1.1.0
idna==3.4
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
model-bakery==1.9.0
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==22.0
pluggy==1.0.0
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
pytest==7.2.0
pytest-django==4.5.2
python3-openid==3.2.0
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
simplejson==3.18.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.13
